+++ src/util/types.js
@@ -207,6 +207,17 @@ propTypes.ownListing = shape({
   images: arrayOf(propTypes.image),
 });
 
+export const BOOKING_STATE_PENDING = 'pending';
+export const BOOKING_STATE_ACCEPTED = 'accepted';
+export const BOOKING_STATE_DECLINED = 'declined';
+export const BOOKING_STATE_CANCELLED = 'cancelled';
+export const BOOKING_STATES = [
+  BOOKING_STATE_PENDING,
+  BOOKING_STATE_ACCEPTED,
+  BOOKING_STATE_DECLINED,
+  BOOKING_STATE_CANCELLED,
+];
+
 // Denormalised booking object
 propTypes.booking = shape({
   id: propTypes.uuid.isRequired,
@@ -214,6 +225,9 @@ propTypes.booking = shape({
   attributes: shape({
     end: instanceOf(Date).isRequired,
     start: instanceOf(Date).isRequired,
+    displayStart: instanceOf(Date),
+    displayEnd: instanceOf(Date),
+    state: oneOf(BOOKING_STATES),
   }),
 });